+++ R/listVeupathdb.R
@@ -2,8 +2,7 @@
 #'
 #' A convenience function to quickly fetch table of genomes and their associated metadata from VEupathDB.
 #'
-#' @import readr
-#' @importFrom glue glue
+#' @import readr jsonlite
 #' @export
 #'
 #' @param customFields A vector of custom fields desired to be fetched. "primary_key" is mandatory field. Other fields can be supplied and can be chosen from (but are not limited to): "annotation_source", "annotation_version", "arraygenecount", "chipchipgenecount", "chromosomeCount", "codinggenecount", "communitycount", "contigCount", "ecnumbercount", "estcount", "genecount", "genecount_number", "genome_source", "genome_version", "gocount", "is_in_apollo", "is_reference_strain", "megabps", "ncbi_tax_id", "ncbi_taxon_url", "organism", "organism_full", "orthologcount", "othergenecount", "popsetcount", "project_id", "proteomicscount", "pseudogenecount", "rnaseqcount", "rtpcrcount", "snpcount", "species", "species_ncbi_tax_id", "species_ncbi_taxon_url", "supercontigCount", "tfbscount", "URLcdsFasta", "URLGenomeFasta", "URLgff", "URLproteinFasta", "URLtranscriptFasta". For more fields, refer to the VEuPathDB Documentation
@@ -16,32 +15,42 @@
 #' }
 #'
 
-
-listVeupathdb <- function(customFields=NULL){
-
-  if(is.null(customFields)){
-    df <- readr::read_tsv(
-      utils::URLencode(
-        'https://veupathdb.org/veupathdb/service/record-types/organism/searches/GenomeDataTypes/reports/attributesTabular?reportConfig={"attributes": ["primary_key","species","URLGenomeFasta","URLcdsFasta","URLtranscriptFasta","URLproteinFasta","project_id","genecount","URLgff","project_id","URLgff","project_id","genome_source","annotation_source"],"includeHeader":true,"attachmentType":"plain"}'
-        ),progress = FALSE,show_col_types = FALSE
-      )
-    return(df)
-
-
-  } else{
-    # Collapse into comma-separated string of quoted field names
-    field_str <- glue_collapse(glue('"{customFields}"'), sep = ",")
-
-    # Build full URL with correct JSON query syntax
-    url <- glue(
-      'https://veupathdb.org/veupathdb/service/record-types/organism/searches/GenomeDataTypes/reports/attributesTabular?reportConfig={{"attributes": [{field_str}],"includeHeader":true,"attachmentType":"plain"}}'
-    )
-
-    df <- readr::read_tsv(utils::URLencode(url),progress = FALSE,show_col_types = FALSE)
-    return(df)
-  }
-
-
-
+listVeupathdb <- function(customFields = NULL) {
+  
+  defaultFields <- c(
+    "primary_key",
+    "species",
+    "URLGenomeFasta",
+    "URLcdsFasta",
+    "URLtranscriptFasta",
+    "URLproteinFasta",
+    "project_id",
+    "genecount",
+    "URLgff",
+    "genome_source",
+    "annotation_source"
+  )
+  
+  fields <- if (is.null(customFields)) defaultFields else customFields
+  
+  report_config <- jsonlite::toJSON(
+    list(
+      attributes = fields,
+      includeHeader = jsonlite::unbox(TRUE),
+      attachmentType = jsonlite::unbox("plain")
+    ),
+    auto_unbox = FALSE
+  )
+  
+  url <- paste0(
+    "https://veupathdb.org/veupathdb/service/record-types/organism/searches/GenomeDataTypes/reports/attributesTabular",
+    "?reportConfig=",
+    utils::URLencode(report_config, reserved = TRUE)
+  )
+  
+  readr::read_tsv(
+    url,
+    progress = FALSE,
+    show_col_types = FALSE
+  )
 }
-