+++ NEWS.md
@@ -1,4 +1,15 @@
-# plasmoRUtils 1.1.0
+# plasmoRUtils 1.1.1
+
+## Changes in version 1.1.1
+
+- Improved error handling in `searchTedConsensus` function for marginal cases where uniprot IDs throws 404 error.
+- Error handling improved in `easyAF2Signal` for marginal cases like `A0A140KXW6`.
+- Deprecation of `searchGSC` due to compliance issues. This function has now been replaced with `searchRS` to query Research Square programmatically using its API.
+- Removed `ggpubr`, `ggsci` and `glue` to slim down package.
+- New function: `getPreconfiguredTableOrthomcl()` is a new function to fetch preconfigured tables from OrthoMCL. It can be used to map old OG IDs to new OG IDs, retrive tables of PFam Architecture of Each Protein, Summary of Pfam domains, Summary of EC Numbers etc.
+- Fixed latency in `searchHP` by using `httr2`.
+- Revised `searchPM` function by replacing depreciated functions with new functions. 
+- Revised `searchPhPl` to handle borderline cases.
 
 
 ## Changes in version 1.1.0